Single mantis directory (#1061)
* single mantis dir, no recopying

* post clustering mask suffix

* generic clustering adjustments

* mask prefix

* remove mask prefix

+++ src/ark/utils/plot_utils.py
@@ -596,7 +596,8 @@ def create_mantis_dir(fovs: List[str], mantis_project_path: Union[str, pathlib.P
                       cluster_type='pixel',
                       mask_suffix: str = "_mask",
                       seg_suffix_name: Optional[str] = "_whole_cell.tiff",
-                      img_sub_folder: str = None):
+                      img_sub_folder: str = None,
+                      new_mask_suffix: str = None):
     """Creates a mantis project directory so that it can be opened by the mantis viewer.
     Copies fovs, segmentation files, masks, and mapping csv's into a new directory structure.
     Here is how the contents of the mantis project folder will look like.
@@ -647,6 +648,8 @@ def create_mantis_dir(fovs: List[str], mantis_project_path: Union[str, pathlib.P
         img_sub_folder (str, optional):
             The subfolder where the channels exist within the `img_data_path`.
             Defaults to None.
+        new_mask_suffix (str, optional):
+            The new suffix added to the copied mask tiffs.
     """
 
     # verify the type of clustering provided is valid
@@ -672,6 +675,10 @@ def create_mantis_dir(fovs: List[str], mantis_project_path: Union[str, pathlib.P
 
     # Save the segmentation tiff or not
     save_seg_tiff: bool = all(v is not None for v in [seg_dir, seg_suffix_name])
+
+    # if no new suffix specified, copy over with original mask name
+    if not new_mask_suffix:
+        new_mask_suffix = mask_suffix
 
     map_df = map_df.loc[:, [f'{cluster_type}_meta_cluster', f'{cluster_type}_meta_cluster_rename']]
     # remove duplicates from df
@@ -719,17 +726,18 @@ def create_mantis_dir(fovs: List[str], mantis_project_path: Union[str, pathlib.P
         # copy mask into new folder
         mask_name: str = mn + mask_suffix + ".tiff"
         shutil.copy(os.path.join(mask_output_dir, mask_name),
-                    os.path.join(output_dir, 'population{}.tiff'.format(mask_suffix)))
+                    os.path.join(output_dir, 'population{}.tiff'.format(new_mask_suffix)))
 
         # copy the segmentation files into the output directory
         # if `seg_dir` or `seg_name` is none, then skip copying
         if save_seg_tiff:
-            seg_name: str = fov + seg_suffix_name
-            shutil.copy(os.path.join(seg_dir, seg_name),
-                        os.path.join(output_dir, 'cell_segmentation.tiff'))
+            if not os.path.exists(os.path.join(output_dir, 'cell_segmentation.tiff')):
+                seg_name: str = fov + seg_suffix_name
+                shutil.copy(os.path.join(seg_dir, seg_name),
+                            os.path.join(output_dir, 'cell_segmentation.tiff'))
 
         # copy mapping into directory
-        map_df.to_csv(os.path.join(output_dir, 'population{}.csv'.format(mask_suffix)),
+        map_df.to_csv(os.path.join(output_dir, 'population{}.csv'.format(new_mask_suffix)),
                       index=False)
